backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0

# Browser Automation
playwright==1.40.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Logging
loguru==0.7.2

# Task Queue (Optional)
celery==5.3.4
redis==5.0.1

# Email (Optional)
python-emails==0.6.0

psutil-7.0.0
flask 
flask-jwt-extended

pymongo
aiofiles